--- OpenGLEX_3_Using_Original/OpenGLEX_3_Using_Original.pro
@@ -1,23 +1,0 @@
-#-------------------------------------------------
-#
-# Project created by QtCreator 2016-02-11T07:09:05
-#
-#-------------------------------------------------
-
-QT       += core gui
-
-greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
-
-TARGET = OpenGLEX_3_Using_Original
-TEMPLATE = app
-
-
-SOURCES += main.cpp\
-        mainwindow.cpp \
-    shadercode.cpp \
-    xiaoglwidget.cpp
-
-HEADERS  += mainwindow.h \
-    xiaoglwidget.h
-
-FORMS    += mainwindow.ui
