--- s1.dot/myQGLWidget.cpp
@@ -1,155 +1,0 @@
-#include <GL/glew.h>
-#include <QtGui>
-#include <QtOpenGL>
-#include <glut.h>
-#include "myQGLWidget.h"
-
-
-MyQGLWidget::MyQGLWidget(QWidget *parent)
-    : QGLWidget(parent)
-{
-    setFormat(QGLFormat(QGL::DoubleBuffer | QGL::DepthBuffer));
-}
-
-void MyQGLWidget::initializeGL()
-{
-    qglClearColor(Qt::blue);  //设置清除屏幕所用的颜色
-    //glClearColor(1.0f, 0.0f, 0.0f, 0.0f);
-    glShadeModel(GL_FLAT);
-    glEnable(GL_DEPTH_TEST);  // 启用深度测试
-    glEnable(GL_CULL_FACE);   // 所作深度测试的类型
-}
-
-void MyQGLWidget::resizeGL(int width, int height)
-{
-    glViewport(0, 0, width, height);
-    glMatrixMode(GL_PROJECTION);               //选择投影矩阵
-    glLoadIdentity();                          //重置投影矩阵
-    GLfloat x = GLfloat(width) / height;
-    glFrustum(-x, +x, -1.0, +1.0, 4.0, 15.0);
-    glMatrixMode(GL_MODELVIEW);                // 选择模型观察矩阵
-}
-
-void MyQGLWidget::paintGL()
-{
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-    draw();
-}
-
-void MyQGLWidget::draw()
-{
-//    // Build and compile our shader program
-//    // Vertex shader
-//    GLuint vertexShader = glCreateShader(GL_VERTEX_SHADER);
-//    glShaderSource(vertexShader, 1, &vertexShaderSource, NULL);
-//    glCompileShader(vertexShader);
-//    // Check for compile time errors
-//    GLint success;
-//    GLchar infoLog[512];
-//    glGetShaderiv(vertexShader, GL_COMPILE_STATUS, &success);
-//    if (!success)
-//    {
-//        glGetShaderInfoLog(vertexShader, 512, NULL, infoLog);
-//        std::cout << "ERROR::SHADER::VERTEX::COMPILATION_FAILED\n" << infoLog << std::endl;
-//    }
-//    // Fragment shader
-//    GLuint fragmentShader = glCreateShader(GL_FRAGMENT_SHADER);
-//    glShaderSource(fragmentShader, 1, &fragmentShaderSource, NULL);
-//    glCompileShader(fragmentShader);
-//    // Check for compile time errors
-//    glGetShaderiv(fragmentShader, GL_COMPILE_STATUS, &success);
-//    if (!success)
-//    {
-//        glGetShaderInfoLog(fragmentShader, 512, NULL, infoLog);
-//        std::cout << "ERROR::SHADER::FRAGMENT::COMPILATION_FAILED\n" << infoLog << std::endl;
-//    }
-//    // Link shaders
-//    GLuint shaderProgram = glCreateProgram();
-//    glAttachShader(shaderProgram, vertexShader);
-//    glAttachShader(shaderProgram, fragmentShader);
-//    glLinkProgram(shaderProgram);
-//    // Check for linking errors
-//    glGetProgramiv(shaderProgram, GL_LINK_STATUS, &success);
-//    if (!success) {
-//        glGetProgramInfoLog(shaderProgram, 512, NULL, infoLog);
-//        std::cout << "ERROR::SHADER::PROGRAM::LINKING_FAILED\n" << infoLog << std::endl;
-//    }
-//    glDeleteShader(vertexShader);
-//    glDeleteShader(fragmentShader);
-
-
-//    // Set up vertex data (and buffer(s)) and attribute pointers
-//    //GLfloat vertices[] = {
-//    //  // First triangle
-//    //   0.5f,  0.5f,  // Top Right
-//    //   0.5f, -0.5f,  // Bottom Right
-//    //  -0.5f,  0.5f,  // Top Left
-//    //  // Second triangle
-//    //   0.5f, -0.5f,  // Bottom Right
-//    //  -0.5f, -0.5f,  // Bottom Left
-//    //  -0.5f,  0.5f   // Top Left
-//    //};
-//    GLfloat vertices[] = {
-//         0.5f,  0.5f, 0.0f,  // Top Right
-//         0.5f, -0.5f, 0.0f,  // Bottom Right
-//        -0.5f, -0.5f, 0.0f,  // Bottom Left
-//        -0.5f,  0.5f, 0.0f   // Top Left
-//    };
-//    GLuint indices[] = {  // Note that we start from 0!
-//        0, 1, 3,  // First Triangle
-//        1, 2, 3   // Second Triangle
-//    };
-//    GLuint VBO, VAO, EBO;
-//    glGenVertexArrays(1, &VAO);
-//    glGenBuffers(1, &VBO);
-//    glGenBuffers(1, &EBO);
-//    // Bind the Vertex Array Object first, then bind and set vertex buffer(s) and attribute pointer(s).
-//    glBindVertexArray(VAO);
-
-//    glBindBuffer(GL_ARRAY_BUFFER, VBO);
-//    glBufferData(GL_ARRAY_BUFFER, sizeof(vertices), vertices, GL_STATIC_DRAW);
-
-//    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, EBO);
-//    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(indices), indices, GL_STATIC_DRAW);
-
-//    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(GLfloat), (GLvoid*)0);
-//    glEnableVertexAttribArray(0);
-
-//    glBindBuffer(GL_ARRAY_BUFFER, 0); // Note that this is allowed, the call to glVertexAttribPointer registered VBO as the currently bound vertex buffer object so afterwards we can safely unbind
-
-//    glBindVertexArray(0); // Unbind VAO (it's always a good thing to unbind any buffer/array to prevent strange bugs), remember: do NOT unbind the EBO, keep it bound to this VAO
-
-
-//    // Uncommenting this call will result in wireframe polygons.
-//    //glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
-
-//    // Game loop
-//    while (!glfwWindowShouldClose(window))
-//    {
-//        // Check if any events have been activiated (key pressed, mouse moved etc.) and call corresponding response functions
-//        glfwPollEvents();
-
-//        // Render
-//        // Clear the colorbuffer
-//        glClearColor(0.2f, 0.3f, 0.3f, 1.0f);
-//        glClear(GL_COLOR_BUFFER_BIT);
-
-//        // Draw our first triangle
-//        glUseProgram(shaderProgram);
-//        glBindVertexArray(VAO);
-//        //glDrawArrays(GL_TRIANGLES, 0, 6);
-//        glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, 0);
-//        glBindVertexArray(0);
-
-//        // Swap the screen buffers
-//        glfwSwapBuffers(window);
-//    }
-//    // Properly de-allocate all resources once they've outlived their purpose
-//    glDeleteVertexArrays(1, &VAO);
-//    glDeleteBuffers(1, &VBO);
-//    glDeleteBuffers(1, &EBO);
-//    // Terminate GLFW, clearing any resources allocated by GLFW.
-//    glfwTerminate();
-//    return 0;
-
-}
