+++ s1.openGLWindow/myQGLWidget.h
@@ -1,9 +1,10 @@
 #ifndef MY_QGLWIDGET_H
 #define MY_QGLWIDGET_H
 
-#include <QGLWidget>
+#include <QOpenGLWidget>
+#include <QOpenGLFunctions>
 
-class MyQGLWidget : public QGLWidget
+class MyQGLWidget : public QOpenGLWidget , protected QOpenGLFunctions
 {
     Q_OBJECT
 
@@ -14,9 +15,6 @@ protected:
     void initializeGL();
     void resizeGL(int width, int height);
     void paintGL();
-
-private:
-    void draw();
 
 };
