+++ s1.openGLWindow/openGLWindow.pro
@@ -1,5 +1,7 @@
 TEMPLATE      = app
-QT           += gui
+QT            += core gui
+QT            += opengl
+greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
 CONFIG       += console
 HEADERS       = myQGLWidget.h
 SOURCES       = main.cpp \
